\name{schisto}
\docType{data}
\alias{schisto}
\title{Schistosomiasis Data Set}
\description{Data on types of water exposure and schistosomiasis infection.}
\usage{data(schisto)}
\format{
A data frame containing the outcome variable (stoolpos -- 1 = positive stool sample (infected), 0 = negative stool sample), 7 water contact exposure variables (total time subject spent doing each activity: laundry, tool washing, bathing, swimming, ditch digging, rice planting, fishing), and two categorical covariates (village id number and age category).

Note: Village label does not correspond to the id numbers from the paper.
}
\source{Spear RC, Seto E, Liang S, Birkner M, Hubbard A, Qiu D, Yang C, Zhong B, Xu F, Gu X, Davis GM (2004). "Factors Influencing the Transmission of Schistosoma Japonicum in the Mountains of Sichuan Province of China." The American Journal of Tropical Medicine and Hygiene, 70(1), 48-56.}
\keyword{datasets}
