\name{wcgs}
\docType{data}
\alias{wcgs}
\title{Subset of Data from Western Collaborative Group Study}
\description{A subset of the data from the Western Collaborative Group Study (described in Rosenman et al. 1975)}
\usage{data(wcgs)}
\format{
A data frame containing the following variables (all are type integer except for bmi, which is double):
\itemize{
  \item chd69: Outcome variable -- did subject get CHD by end of follow up (1969)?
  \item sbp0: Systolic blood pressure at baseline.
  \item dbp0: Diastolic blood pressure at baseline.
  \item chol0: Fasting serum cholesterol at baseline (contains 12 missing values -- NA).
  \item ncigs0: Cigarettes per day.
  \item dibpat0: Behavior type -- 1 = type A, 2 = type B.
  \item age0: Age in years at baseline.
  \item height0: Height in inches.
  \item weight0: Weight in lbs.
  \item bmi0: BMI based on height0 and weight0.
}
}
\source{Rosenman RH, Brand RJ, Jenkins CD, Friedman M, Straus R, Wurm M (1975). "Coronary Heart Disease in the Western Collaborative Group Study. Final Follow-up Experience of 8 1/2 Years." JAMA, 233(8), 872-7.}
\keyword{datasets}
